+++ R/vein-deprecated.R
@@ -1,0 +1,9 @@
+## vein-deprecated.r
+#' @title Deprecated functions in package \pkg{vein}.
+#' @description The functions listed below are deprecated and will be defunct in
+#'   the near future. When possible, alternative functions with similar
+#'   functionality are also mentioned. Help pages for deprecated functions are
+#'   available at \code{help("-deprecated")}.
+#' @name vein-deprecated
+#' @keywords internal
+NULL